requirements.txt
streamlit==1.31.0
youtube-transcript-api==0.6.2
google-generativeai==0.3.2
